+++ src/api/gateway.go
@@ -62,4 +62,8 @@ type Gatewayer interface {
 	GetHealth() (*daemon.Health, error)
 	UnloadWallet(id string) error
 	VerifyTxnVerbose(txn *coin.Transaction) ([]wallet.UxBalance, bool, error)
+	GetAllNotes() []notes.Note
+	GetNoteByTxID(txID string) notes.Note
+	AddNote(notes notes.Note) error
+	RemoveNote(txID string) error
 }