+++ src/daemon/gateway_test.go
@@ -11,86 +11,7 @@ import (
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/coin"
-	"github.com/skycoin/skycoin/src/testutil"
-	"github.com/skycoin/skycoin/src/util/collections"
 )
-
-func TestFbyAddresses(t *testing.T) {
-	uxs := make(coin.UxArray, 5)
-	addrs := make([]cipher.Address, 5)
-	for i := 0; i < 5; i++ {
-		addrs[i] = testutil.MakeAddress()
-		uxs[i] = coin.UxOut{
-			Body: coin.UxBody{
-				Address: addrs[i],
-			},
-		}
-	}
-
-	tests := []struct {
-		name    string
-		addrs   []string
-		outputs []coin.UxOut
-		want    []coin.UxOut
-	}{
-		// TODO: Add test cases.
-		{
-			"filter with one address",
-			[]string{addrs[0].String()},
-			uxs[:2],
-			uxs[:1],
-		},
-		{
-			"filter with multiple addresses",
-			[]string{addrs[0].String(), addrs[1].String()},
-			uxs[:3],
-			uxs[:2],
-		},
-	}
-	for _, tt := range tests {
-		// fmt.Printf("want:%+v\n", tt.want)
-		outs := FbyAddresses(tt.addrs)(tt.outputs)
-		require.Equal(t, outs, coin.UxArray(tt.want))
-	}
-}
-
-func TestFbyHashes(t *testing.T) {
-	uxs := make(coin.UxArray, 5)
-	addrs := make([]cipher.Address, 5)
-	for i := 0; i < 5; i++ {
-		addrs[i] = testutil.MakeAddress()
-		uxs[i] = coin.UxOut{
-			Body: coin.UxBody{
-				Address: addrs[i],
-			},
-		}
-	}
-
-	tests := []struct {
-		name    string
-		hashes  []string
-		outputs coin.UxArray
-		want    coin.UxArray
-	}{
-		// TODO: Add test cases.
-		{
-			"filter with one hash",
-			[]string{uxs[0].Hash().Hex()},
-			uxs[:2],
-			uxs[:1],
-		},
-		{
-			"filter with multiple hash",
-			[]string{uxs[0].Hash().Hex(), uxs[1].Hash().Hex()},
-			uxs[:3],
-			uxs[:2],
-		},
-	}
-	for _, tt := range tests {
-		outs := FbyHashes(tt.hashes)(tt.outputs)
-		require.Equal(t, outs, tt.want)
-	}
-}
 
 func TestGateway_GetWalletDir(t *testing.T) {
 	tests := []struct {
@@ -313,18 +234,27 @@ func TestGateway_ReloadWallets(t *testing.T) {
 
 func TestGateway_Spend(t *testing.T) {
 	tests := []struct {
-		name            string
-		enableWalletAPI bool
-		walletID        string
-		coins           uint64
-		dest            cipher.Address
-		result          *coin.Transaction
-		err             error
+		name              string
+		enableWalletAPI   bool
+		enableSpendMethod bool
+		walletID          string
+		coins             uint64
+		dest              cipher.Address
+		result            *coin.Transaction
+		err               error
 	}{
 		{
-			name:            "wallet api disabled",
-			enableWalletAPI: false,
-			err:             wallet.ErrWalletAPIDisabled,
+			name:              "wallet api disabled",
+			enableWalletAPI:   false,
+			enableSpendMethod: true,
+			err:               wallet.ErrWalletAPIDisabled,
+		},
+
+		{
+			name:              "spend method disabled",
+			enableWalletAPI:   true,
+			enableSpendMethod: false,
+			err:               ErrSpendMethodDisabled,
 		},
 	}
 
@@ -332,7 +262,8 @@ func TestGateway_Spend(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			gw := &Gateway{
 				Config: GatewayConfig{
-					EnableWalletAPI: tc.enableWalletAPI,
+					EnableWalletAPI:   tc.enableWalletAPI,
+					EnableSpendMethod: tc.enableSpendMethod,
 				},
 			}
 			res, err := gw.Spend(tc.walletID, nil, tc.coins, tc.dest)
@@ -430,14 +361,9 @@ func TestGateway_CreateTransaction(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			// Assume all API sets enabled by default
-			enabledAPISets := collections.NewStringSet("READ")
-			if tc.enableWalletAPI {
-				_ = enabledAPISets.Set("WALLET") // nolint errchec // nolint errcheckk
-			}
 			gw := &Gateway{
 				Config: GatewayConfig{
-					EnabledAPISets: enabledAPISets,
+					EnableWalletAPI: tc.enableWalletAPI,
 				},
 			}